Split sockets into modules.
- Create separate modules for message and contact sockets
